refactor(native-filters): Refactor filters config modal module (#13268)
* refactor(native-filters): Refactor filters config modal module

* fix: fix import

* lint: fix import
