+++ superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/state.ts
@@ -16,65 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { useEffect, useMemo } from 'react';
-import { useSelector } from 'react-redux';
-import { t } from '@superset-ui/core';
-import { Charts, Layout, RootState } from 'src/dashboard/types';
-import { DASHBOARD_ROOT_ID } from 'src/dashboard/util/constants';
-import {
-  CHART_TYPE,
-  DASHBOARD_ROOT_TYPE,
-} from 'src/dashboard/util/componentTypes';
+import { useEffect } from 'react';
 import { FormInstance } from 'antd/lib/form';
 import { getChartDataRequest } from 'src/chart/chartAction';
-import { NativeFilterState } from 'src/dashboard/reducers/types';
-import { NativeFiltersForm, TreeItem } from './types';
-import { buildTree, setFilterFieldValues, useForceUpdate } from './utils';
-import { Filter } from '../types';
-import { getFormData } from '../utils';
-
-export function useFiltersState() {
-  return useSelector<any, NativeFilterState>(
-    state => state.nativeFilters.filtersState,
-  );
-}
-
-export function useFilterScopeTree(): {
-  treeData: [TreeItem];
-  layout: Layout;
-} {
-  const layout = useSelector<RootState, Layout>(
-    ({ dashboardLayout: { present } }) => present,
-  );
-
-  const charts = useSelector<RootState, Charts>(({ charts }) => charts);
-  const tree = {
-    children: [],
-    key: DASHBOARD_ROOT_ID,
-    type: DASHBOARD_ROOT_TYPE,
-    title: t('All panels'),
-  };
-
-  // We need to get only nodes that have charts as children or grandchildren
-  const validNodes = useMemo(
-    () =>
-      Object.values(layout).reduce<string[]>((acc, cur) => {
-        if (cur?.type === CHART_TYPE) {
-          return [...new Set([...acc, ...cur?.parents, cur.id])];
-        }
-        return acc;
-      }, []),
-    [layout],
-  );
-
-  useMemo(() => {
-    buildTree(layout[DASHBOARD_ROOT_ID], tree, layout, charts, validNodes);
-  }, [charts, layout, tree]);
-
-  return { treeData: [tree], layout };
-}
+import { NativeFiltersForm } from '../types';
+import { setFilterFieldValues, useForceUpdate } from './utils';
+import { Filter } from '../../types';
+import { getFormData } from '../../utils';
 
 // When some fields in form changed we need re-fetch data for Filter defaultValue
+// eslint-disable-next-line import/prefer-default-export
 export const useBackendFormUpdate = (
   form: FormInstance<NativeFiltersForm>,
   filterId: string,
